Refactor: Create, modify & rename components
