+++ packages/frontend/src/components/dynamic/album/AlbumLoading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const AlbumLoading = () => {
+  return <div>AlbumLoading</div>;
+};
+
+export default AlbumLoading;
