+++ packages/frontend/src/components/dynamic/band/BandLoading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const BandLoading = () => {
+  return <div>BandLoading</div>;
+};
+
+export default BandLoading;
