+++ packages/frontend/src/components/dynamic/dynamic-layout/DynamicHeader.tsx
@@ -3,18 +3,18 @@ import Link from "next/link";
 import React from "react";
 import { IoPerson } from "react-icons/io5";
 
-interface headerDataProps {
+interface dynamicHeaderDataProps {
   link: string;
   icon: React.ReactNode;
 }
 
 const iconStyle = "h-[20px] w-[20px]";
 
-const headerData: headerDataProps[] = [
+const dynamicHeaderData: dynamicHeaderDataProps[] = [
   { link: "", icon: <IoPerson className={iconStyle} /> },
 ];
 
-const Header = () => {
+const DynamicHeader = () => {
   return (
     <nav className="fixed top-0 z-50 flex h-[70px] w-full items-center justify-between border bg-white p-[10px] px-5">
       <Link href={"/"} className="flex items-center justify-center">
@@ -24,7 +24,7 @@ const Header = () => {
         </h1>
       </Link>
       <div className="flex items-center gap-3">
-        {headerData.map((item, i) => (
+        {dynamicHeaderData.map((item, i) => (
           <Link key={i} href={item.link} className="rounded border p-1">
             {item.icon}
           </Link>
@@ -34,4 +34,4 @@ const Header = () => {
   );
 };
 
-export default Header;
+export default DynamicHeader;
