+++ packages/frontend/src/components/dynamic/song/SongKeyTransposition.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const SongKeyTransposition = () => {
+  return (
+    <>
+      {/* TODO: transpose key section */}
+      <div className="rounded border-2 p-5"></div>
+    </>
+  );
+};
+
+export default SongKeyTransposition;
