+++ packages/frontend/src/components/dynamic/song/SongLoading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SongLoading = () => {
+  return <div>SongLoading</div>;
+};
+
+export default SongLoading;
