+++ packages/frontend/src/components/index/IndexFooter.tsx
@@ -1,28 +1,28 @@
 import React from "react";
 import Link from "next/link";
 
-const leftFooterData = [
+const leftIndexFooterData = [
   { link: "/", text: "About" },
   //   { link: "/", text: "" },
 ];
 
-const rightFooterData = [
+const rightIndexFooterData = [
   { link: "/", text: "Language" },
   { link: "/", text: "Privacy" },
 ];
 
-const Footer = () => {
+const IndexFooter = () => {
   return (
     <div className="fixed bottom-0 flex h-12 w-full items-center justify-between border px-3">
       <div className="flex gap-3 text-xs text-neutral-500">
-        {leftFooterData.map((item, i) => (
+        {leftIndexFooterData.map((item, i) => (
           <Link key={i} href={item.link} className="hover:text-neutral-400">
             {item.text}
           </Link>
         ))}
       </div>
       <div className="flex gap-3 text-xs text-neutral-500">
-        {rightFooterData.map((item, i) => (
+        {rightIndexFooterData.map((item, i) => (
           <Link key={i} href={item.link} className="hover:text-neutral-400">
             {item.text}
           </Link>
@@ -32,4 +32,4 @@ const Footer = () => {
   );
 };
 
-export default Footer;
+export default IndexFooter;
