+++ packages/frontend/src/components/search/SearchAlbumList.tsx
@@ -1,0 +1,58 @@
+import Link from "next/link";
+import React, { useState } from "react";
+import { type Song } from "@prisma/client";
+
+interface SearchAlbumListProps {
+  showBand: boolean | undefined;
+  filteredSongList: Song[];
+}
+
+const SearchAlbumList: React.FC<SearchAlbumListProps> = ({
+  showBand,
+  filteredSongList,
+}) => {
+  const [activeList, setActiveList] = useState(-1);
+  return (
+    <>
+      {showBand && (
+        <div className="hidden h-fit w-5/12 flex-col gap-3 lg:flex">
+          <h1 className="text-lg font-semibold">
+            Albums from {filteredSongList[0]?.original_band}
+          </h1>
+          {[
+            ...new Set(
+              filteredSongList
+                .filter((items) => items.album)
+                .map((items) => items.album),
+            ),
+          ].map((album, i) => (
+            <Link
+              href={`/album/${album?.toLowerCase().replace(/ /g, "-")}`}
+              key={i}
+              className={`${
+                activeList === i ? "bg-[#f5f5f6] shadow-md" : ""
+              } rounded border-2 p-3 px-5`}
+              onMouseEnter={() => setActiveList(i)}
+              onMouseLeave={() => setActiveList(-1)}
+            >
+              <h1 className="font-semibold">{album}</h1>
+              <p className="text-sm text-slate-500">
+                {
+                  filteredSongList.filter((items) => {
+                    return (
+                      items.album?.toLowerCase().replace(/ /g, "") ===
+                      album?.toLowerCase().replace(/ /g, "")
+                    );
+                  }).length
+                }{" "}
+                songs
+              </p>
+            </Link>
+          ))}
+        </div>
+      )}
+    </>
+  );
+};
+
+export default SearchAlbumList;
