+++ packages/frontend/src/components/search/SearchLyrics.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SearchLyrics = () => {
+  return <></>;
+};
+
+export default SearchLyrics;
