+++ packages/frontend/src/pages/song/[song].tsx
@@ -1,45 +1,15 @@
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import React, { useState, type ReactElement, useEffect, useRef } from "react";
-import Layout from "@/components/dynamic/layout/Layout";
-import Link from "next/link";
+import DynamicLayout from "@/components/dynamic/dynamic-layout/DynamicLayout";
 import { useRouter } from "next/router";
-import markdownit from "markdown-it";
-// @ts-expect-error no declaration file
-import chords from "markdown-it-chords";
-import SongBreadcrumb from "@/components/dynamic/SongBreadcrumb";
+import SongBreadcrumb from "@/components/dynamic/song/SongBreadcrumb";
 import Head from "next/head";
+import { type Song } from "@prisma/client";
+import SongDetails from "@/components/dynamic/song/SongDetails";
+import SongKeyTransposition from "@/components/dynamic/song/SongKeyTransposition";
+import SongLyrics from "@/components/dynamic/song/SongLyrics";
+import SongLoading from "@/components/dynamic/song/SongLoading";
 
-type Song = {
-  id?: string;
-  name: string;
-  alt_name?: string;
-  song_language: string;
-  original_key: string;
-  original_band: string;
-  album?: string;
-  original_youtube_url?: string;
-  chord_lyrics: string;
-  main_key_link?: string;
-  sub_key_link?: string;
-  eg_link?: string;
-  ag_link?: string;
-  bass_link?: string;
-  drum_link?: string;
-  tags?: string[];
-  sequencer_files?: string[];
-  sub_voice_file?: string;
-};
-
-const md = markdownit({
-  breaks: true,
-});
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-md.use(chords);
-
-const Song = () => {
+const DynamicSong = () => {
   const router = useRouter();
   const [songList, setSongList] = useState<Song[]>([]);
   const [filteredSongList, setFilteredSongList] = useState<Song[]>([]);
@@ -69,7 +39,7 @@ const Song = () => {
     const filteredSongList = songList.filter(
       (items) =>
         router.query.song &&
-        `${items.name.toLowerCase().replace(/ /g, "-")}` ===
+        `${items.name!.toLowerCase().replace(/ /g, "-")}` ===
           router.query.song.toString(),
     );
     setFilteredSongList(filteredSongList);
@@ -80,7 +50,9 @@ const Song = () => {
     return params.get("v");
   };
 
-  if (!isLoading) {
+  if (isLoading) {
+    <SongLoading />;
+  } else {
     return (
       <>
         <Head>
@@ -94,85 +66,21 @@ const Song = () => {
           return (
             <div key={i} className="flex flex-col gap-5">
               <SongBreadcrumb
-                name={items.name}
+                name={items.name!}
                 album={items.album}
-                original_band={items.original_band}
+                original_band={items.original_band!}
               />
               <div className="flex flex-col gap-5 pb-5 md:flex-row">
-                {/* left */}
-                <div className="flex h-fit flex-col rounded border-2 p-5">
-                  <div className="h-[150px] w-full md:w-[200px] lg:h-[150px] lg:w-[300px]">
-                    <iframe
-                      src={embedUrl}
-                      allowFullScreen
-                      className="me-5 h-full w-full rounded md:me-0"
-                    ></iframe>
-                  </div>
-                  <div>
-                    <div className="flex flex-col gap-2 py-3">
-                      <p className="flex flex-col truncate">
-                        <span className="font-semibold">Song Name</span>
-                        <span className="text-sm text-neutral-500">
-                          {items.name}{" "}
-                          {items.alt_name &&
-                            items.alt_name.trim() !== "-" &&
-                            items.alt_name}
-                        </span>
-                      </p>
-                      {items.album && (
-                        <p className="flex flex-col truncate">
-                          <span className="font-semibold">Album</span>
-                          <Link
-                            href={`/album/${items.album
-                              .toLowerCase()
-                              .replace(/ /g, "-")}`}
-                            className="w-fit text-sm text-neutral-500 underline md:no-underline md:hover:underline"
-                          >
-                            {items.album}
-                          </Link>
-                        </p>
-                      )}
-                      <p className="flex flex-col truncate">
-                        <span className="font-semibold">Band</span>
-                        <Link
-                          href={`/band/${items.original_band
-                            .toLowerCase()
-                            .replace(/ /g, "-")}`}
-                          className="w-fit text-sm text-neutral-500 underline md:no-underline md:hover:underline"
-                        >
-                          {items.original_band}
-                        </Link>
-                      </p>
-                    </div>
-                    <hr />
-                    <div className="flex flex-col gap-2 py-3">
-                      <p className="flex flex-col truncate">
-                        <span className="font-semibold">Original Key</span>
-                        <span className="text-sm text-neutral-500">
-                          {items.original_key} Major
-                        </span>
-                      </p>
-                    </div>
-                  </div>
-                  {/* TODO: tags */}
-                </div>
-                {/* right */}
+                <SongDetails embedUrl={embedUrl} items={items} />
                 <div className="flex w-full flex-col gap-5">
                   <h1 className="hidden rounded border-2 px-5 py-3 text-4xl font-semibold md:block">
                     {items.name}
                   </h1>
-                  {/* TODO: transpose key section */}
-                  <div className="rounded border-2 p-5"></div>
-                  {/* TODO: chords & lyrics */}
-                  <div className="rounded border-2 p-5">
-                    <p
-                      ref={lyricsRef}
-                      className="text-neutral-500"
-                      dangerouslySetInnerHTML={{
-                        __html: md.render(items.chord_lyrics),
-                      }}
-                    ></p>
-                  </div>
+                  <SongKeyTransposition />
+                  <SongLyrics
+                    lyricsRef={lyricsRef}
+                    chordLyrics={items.chord_lyrics ? items.chord_lyrics : ""}
+                  />
                 </div>
               </div>
             </div>
@@ -183,8 +91,8 @@ const Song = () => {
   }
 };
 
-export default Song;
+export default DynamicSong;
 
-Song.getLayout = function getLayout(page: ReactElement) {
-  return <Layout>{page}</Layout>;
+DynamicSong.getLayout = function getLayout(page: ReactElement) {
+  return <DynamicLayout>{page}</DynamicLayout>;
 };
